+++ src/aps.h
@@ -6,7 +6,7 @@
 #include "mt_af.h"
 
 /* APS callbacks, registered by proper applications */
-typedef void (*ApsMsgCb)(uint16_t addr, uint16_t cluster, void *data, int len);
+typedef void (*ApsMsgCb)(uint64_t addr, uint16_t cluster, void *data, int len);
 
 /**
  * \brief Intialize the APS layer.
